fix(ci): close tofu lane-coverage gap and split step-structure checks
Why
- Restacking onto main's independent tofu-plan-* step split (single
  pr-dryrun step -> 6 dedicated tofu-plan-* steps) left those steps'
  shared if_changed anchor missing select-main-pipeline-selection.ts,
  which this branch's own earlier commit had added to every other
  PR-side step's closure. ci-lane-coverage.test.ts caught the gap.
- The same restack merged main's native-step branch into
  checkStepStructure, pushing its cognitive complexity to 26 (limit
  15) and validate-pipeline-lib.ts to 508 lines (limit 500); a new
  helper in macos-native-dispatch-watch.ts was already over the
  complexity limit on its own.

What
- Add select-main-pipeline-selection.ts to the shared tofu_plan_if_changed
  YAML anchor so all six tofu-plan-* steps gate on selector changes.
- Split checkStepStructure into checkNativeStepStructure,
  checkKubernetesStepStructure, and checkPathGatedIfChanged, moved to a
  new validate-pipeline-step-structure.ts to keep validate-pipeline-lib.ts
  under the max-lines budget. Extracted the leaf fail/scalar/hasTrimmedLine
  parsing primitives into validate-pipeline-parse.ts so the two files
  don't form an import cycle, and repointed every consumer's import to
  whichever file now owns each symbol (no re-exports, per
  custom-rules/no-re-exports).
- Extracted collectRunningNativeJobs out of parseOtherRunningNativeJobs
  in macos-native-dispatch-watch.ts to bring it under the complexity
  limit.

Verification
- bunx turbo run typecheck lint test --filter=@shepherdjerred/root-scripts
- bun --no-install .buildkite/scripts/validate-pipeline.ts
- bun scripts/jscpd-check.ts
