+++ .buildkite/pipeline.yml
@@ -1831,6 +1831,7 @@ steps:
         - ".buildkite/scripts/prepare-ci-changed-base.ts"
         - ".buildkite/scripts/read-buildkite-handoff.ts"
         - ".buildkite/scripts/select-main-pipeline.ts"
+        - ".buildkite/scripts/select-main-pipeline-selection.ts"
         - ".buildkite/scripts/upload-pipeline.sh"
         - ".buildkite/secret-grants.json"
         - ".buildkite/secret-grants.schema.json"
